requirements.txt
customtkinter==5.2.1
packaging==23.0
pillow>=10.0.0
